feat(qwen35-dense): native dense FP8 attn/FFN dispatch + kernels (Holo-3.1/Ornith) (#215)
Restacked onto #214 — applied cleanly (no conflicts). Adds the dense-FP8 attention/
FFN path Ornith-1.0-35B-FP8 needs.
