--- crates/spark-model/src/layers/qwen3_attention/trait_impl/multi_seq/qkv_seq.rs
@@ -1,227 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-//! Sequential (per-token) Q/K/V projection helpers — the fallback that
-//! supports every weight encoding. Split from qkv.rs to keep files <=500 LoC.
-
-use anyhow::Result;
-
-use crate::layers::ops;
-use crate::layers::qwen3_attention::Qwen3AttentionLayer;
-
-impl Qwen3AttentionLayer {
-    /// Sequential per-token Q projection (handles gated and ungated).
-    #[allow(clippy::too_many_arguments)]
-    pub(super) fn ms_qkv_seq_q(
-        &self,
-        fwd: &crate::layer::ForwardContext<'_>,
-        normed_i: spark_runtime::gpu::DevicePtr,
-        q_out_i: spark_runtime::gpu::DevicePtr,
-        q_proj_dim: u32,
-        q_dim: u32,
-        nq: u32,
-        hd: u32,
-        h: usize,
-        stream: u64,
-    ) -> Result<()> {
-        if self.gated {
-            if let Some(fp8) = self.q_weight.as_ref().and_then(|w| w.as_fp8()) {
-                ops::w8a16_gemv(
-                    fwd.gpu,
-                    self.w8a16_gemv_k,
-                    normed_i,
-                    fp8.weight,
-                    fp8.row_scale,
-                    q_out_i,
-                    q_proj_dim,
-                    h as u32,
-                    stream,
-                )?;
-                ops::deinterleave_qg(
-                    fwd.gpu,
-                    self.deinterleave_qg_k,
-                    q_out_i,
-                    1,
-                    nq,
-                    hd,
-                    q_proj_dim,
-                    stream,
-                )?;
-            } else if let Some(nvfp4) = self.q_weight.as_ref().and_then(|w| w.as_nvfp4()) {
-                ops::w4a16_gemv_qg(
-                    fwd.gpu,
-                    self.w4a16_gemv_qg_k,
-                    normed_i,
-                    nvfp4,
-                    q_out_i,
-                    q_proj_dim,
-                    h as u32,
-                    nq,
-                    hd,
-                    stream,
-                )?;
-            } else {
-                ops::dense_gemv(
-                    fwd.gpu,
-                    self.dense_gemv_k,
-                    normed_i,
-                    &self.attn.q_proj,
-                    q_out_i,
-                    q_proj_dim,
-                    h as u32,
-                    stream,
-                )?;
-                ops::deinterleave_qg(
-                    fwd.gpu,
-                    self.deinterleave_qg_k,
-                    q_out_i,
-                    1,
-                    nq,
-                    hd,
-                    q_proj_dim,
-                    stream,
-                )?;
-            }
-        } else if let Some(fp8) = self.q_weight.as_ref().and_then(|w| w.as_fp8()) {
-            ops::w8a16_gemv(
-                fwd.gpu,
-                self.w8a16_gemv_k,
-                normed_i,
-                fp8.weight,
-                fp8.row_scale,
-                q_out_i,
-                q_dim,
-                h as u32,
-                stream,
-            )?;
-        } else if let Some(nvfp4) = self.q_weight.as_ref().and_then(|w| w.as_nvfp4()) {
-            ops::w4a16_gemv(
-                fwd.gpu,
-                self.w4a16_gemv_k,
-                normed_i,
-                nvfp4,
-                q_out_i,
-                q_dim,
-                h as u32,
-                stream,
-            )?;
-        } else {
-            ops::dense_gemv(
-                fwd.gpu,
-                self.dense_gemv_k,
-                normed_i,
-                &self.attn.q_proj,
-                q_out_i,
-                q_dim,
-                h as u32,
-                stream,
-            )?;
-        }
-        Ok(())
-    }
-
-    /// Sequential per-token K + V projections.
-    #[allow(clippy::too_many_arguments)]
-    pub(super) fn ms_qkv_seq_kv(
-        &self,
-        fwd: &crate::layer::ForwardContext<'_>,
-        normed_i: spark_runtime::gpu::DevicePtr,
-        k_out_i: spark_runtime::gpu::DevicePtr,
-        v_out_i: spark_runtime::gpu::DevicePtr,
-        nkv: u32,
-        hd: u32,
-        h: usize,
-        stream: u64,
-    ) -> Result<()> {
-        if let (Some(k_fp8), Some(v_fp8)) = (
-            self.k_weight.as_ref().and_then(|w| w.as_fp8()),
-            self.v_weight.as_ref().and_then(|w| w.as_fp8()),
-        ) {
-            ops::w8a16_gemv(
-                fwd.gpu,
-                self.w8a16_gemv_k,
-                normed_i,
-                k_fp8.weight,
-                k_fp8.row_scale,
-                k_out_i,
-                nkv * hd,
-                h as u32,
-                stream,
-            )?;
-            ops::w8a16_gemv(
-                fwd.gpu,
-                self.w8a16_gemv_k,
-                normed_i,
-                v_fp8.weight,
-                v_fp8.row_scale,
-                v_out_i,
-                nkv * hd,
-                h as u32,
-                stream,
-            )?;
-        } else if let (Some(k_fp4), Some(v_fp4)) = (
-            self.k_weight.as_ref().and_then(|w| w.as_nvfp4()),
-            self.v_weight.as_ref().and_then(|w| w.as_nvfp4()),
-        ) {
-            ops::w4a16_gemv_dual(
-                fwd.gpu,
-                self.w4a16_gemv_dual_k,
-                normed_i,
-                k_fp4,
-                k_out_i,
-                v_fp4,
-                v_out_i,
-                nkv * hd,
-                h as u32,
-                stream,
-            )?;
-        } else {
-            if let Some(nvfp4) = self.k_weight.as_ref().and_then(|w| w.as_nvfp4()) {
-                ops::w4a16_gemv(
-                    fwd.gpu,
-                    self.w4a16_gemv_k,
-                    normed_i,
-                    nvfp4,
-                    k_out_i,
-                    nkv * hd,
-                    h as u32,
-                    stream,
-                )?;
-            } else {
-                ops::dense_gemv(
-                    fwd.gpu,
-                    self.dense_gemv_k,
-                    normed_i,
-                    &self.attn.k_proj,
-                    k_out_i,
-                    nkv * hd,
-                    h as u32,
-                    stream,
-                )?;
-            }
-            if let Some(nvfp4) = self.v_weight.as_ref().and_then(|w| w.as_nvfp4()) {
-                ops::w4a16_gemv(
-                    fwd.gpu,
-                    self.w4a16_gemv_k,
-                    normed_i,
-                    nvfp4,
-                    v_out_i,
-                    nkv * hd,
-                    h as u32,
-                    stream,
-                )?;
-            } else {
-                ops::dense_gemv(
-                    fwd.gpu,
-                    self.dense_gemv_k,
-                    normed_i,
-                    &self.attn.v_proj,
-                    v_out_i,
-                    nkv * hd,
-                    h as u32,
-                    stream,
-                )?;
-            }
-        }
-        Ok(())
-    }
-}
