+++ crates/spark-server/src/scheduler/lifecycle.rs
@@ -65,31 +65,7 @@ pub fn finish_sequence(model: &dyn Model, a: &mut ActiveSeq) {
         0.0
     };
     let ttft_ms = a.decode_start.duration_since(a.request_start).as_secs_f64() * 1000.0;
-    // Prefill throughput: prompt tokens actually computed (excluding prefix-cache
-    // hits) over the time-to-first-token (≈ prefill time; includes queue wait
-    // under concurrency). Logged so prefill tok/s is readable straight from the
-    // server, no client-side derivation.
-    let cached = a.cached_prompt_tokens;
-    let prefill_computed = (a.seq.prompt_len as u32).saturating_sub(cached);
-    // Distinguish "no prefill work (prefix-cache hit)" from a real prefill: a
-    // tok/s figure is only meaningful when tokens were actually computed.
-    let prefill_str = if prefill_computed == 0 {
-        format!("prefill: cache hit ({cached} cached)")
-    } else {
-        let prefill_tps = if ttft_ms > 0.0 {
-            prefill_computed as f64 / (ttft_ms / 1000.0)
-        } else {
-            0.0
-        };
-        if cached > 0 {
-            format!("prefill {prefill_computed} tok @ {prefill_tps:.0} tok/s ({cached} cached)")
-        } else {
-            format!("prefill {prefill_computed} tok @ {prefill_tps:.0} tok/s")
-        }
-    };
-    tracing::info!(
-        "Done: {n} tokens ({reason}) {tps:.1} tok/s, TTFT={ttft_ms:.1}ms, {prefill_str}"
-    );
+    tracing::info!("Done: {n} tokens ({reason}) {tps:.1} tok/s, TTFT={ttft_ms:.1}ms");
     // Cache the full sequence (prompt + generated) in the prefix cache.
     // Must happen BEFORE free_sequence() so block indices are still valid.
     // Enables multi-turn sessions to reuse KV cache for prior assistant responses.
